Peripheral, CharacteristicのNativeMethodを呼ぶためのクラスを作成
CBCentralManagerからそれぞれのProxy部分を切り離す。
また、cs側でnative側のクラスを指すときにnativeという単語を使うことにした。わかりやすいから。
そして、Native側のクラスのProxy(代理)のためNativePeripheralProxyと名付けた
