+++ Assets/SamplePlugin/Runtime/CoreBluetooth/NativeCharacteristicProxy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c658040f5226d42cda8d3ef5ede22109
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
