+++ Assets/SamplePlugin/Runtime/CoreBluetooth/NativePeripheralProxy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5a14914a3ca5c4272bbf5ad29f861e44
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
